Issue #2363 Add more vector configuration optons. Change-Id: I1a3e564629b9e4afe93a0a4ceee8e57681aa0644
Former-commit-id: 86cd59135ec421e31b904eb8f481b124f8cf5f4b [formerly 86cd59135ec421e31b904eb8f481b124f8cf5f4b [formerly 860258e49cc9cb766b661d8533be261dd3aae7ac]] [formerly ff06ea1c5af7caa4cc796c5b5cb3dd968474d4c2] [formerly 86cd59135ec421e31b904eb8f481b124f8cf5f4b [formerly 86cd59135ec421e31b904eb8f481b124f8cf5f4b [formerly 860258e49cc9cb766b661d8533be261dd3aae7ac]] [formerly ff06ea1c5af7caa4cc796c5b5cb3dd968474d4c2] [formerly c993020c0b05756ae00be7cb829d806344ab4f27 [formerly c993020c0b05756ae00be7cb829d806344ab4f27 [formerly 20853f733378677864a5bbb08cd27f47a6646f12]] [formerly ff06ea1c5af7caa4cc796c5b5cb3dd968474d4c2 [formerly ff06ea1c5af7caa4cc796c5b5cb3dd968474d4c2 [formerly 8af298f9ef2c1c03e6be690c07829d0569b1dedd]]]]]
Former-commit-id: c993020c0b05756ae00be7cb829d806344ab4f27 [formerly c993020c0b05756ae00be7cb829d806344ab4f27 [formerly 20853f733378677864a5bbb08cd27f47a6646f12]]
Former-commit-id: 7ee3225aeaaa4e8ecdca4e928bc36dd3c1b4ecdf [formerly 562dec9fa86770c8e6df28a1eab04d3b7fcb2b9c] [formerly f4e2b2a7e243478d9379ff0ab2890d0a6c6dcae5 [formerly bf0642450ca21c7e0d6f5eadc3d466a0fba2de7b]] [formerly a22dfa8e54e1919f1bfbbb5e84db553a9f4a1154 [formerly b582a4f24ca6be6abda566cbc19cfa391e2b7398] [formerly 68a35a4c60c0597c7f3ce8dfd515500924a100b7 [formerly 6c7c693581a84d7bce05b1e6f2b2901e4aa5985d]]]
Former-commit-id: 6396e35fe41471d3b3debc1aa794d30ecb8fff2e [formerly f0959ee1bb8564a2b6c405f3e7d2fd3893f002a2] [formerly 9e9834ea9631576bf65dd430b4baf259b8edef1a [formerly baa1ddd596dd6ba7e1ebe0e1cbe78d99802032e7]]
Former-commit-id: d85d65d69cd18b40626ec0f2852363e885f49028 [formerly 938ace6bae87ae7c07ef83a9aae37b822438583f]
Former-commit-id: 8993ff93a55f7a285cbe5fe435aa9f41460823d8

+++ cave/com.raytheon.viz.core.contours/src/com/raytheon/viz/core/contours/rsc/displays/AbstractGriddedDisplay.java
@@ -53,15 +53,17 @@ import com.vividsolutions.jts.geom.Coordinate;
  * <pre>
  * 
  * SOFTWARE HISTORY
- * Date         Ticket#    Engineer    Description
- * ------------ ---------- ----------- --------------------------
- * Apr 23, 2010            bsteffen    Initial creation
- * Aug 07, 2013 2077       bsteffen    Revise pixel size calculations.
- * Aug 27, 2013 2287       randerso    Replaced hard coded constant with densityFactor
- *                                     parameter to allow application specific density
- *                                     scaling to better match A1 displays
- * Sep 10, 2013 DR 16257   MPorricelli  Fix so that wind for global grids displays on
- *                                      mercator maps.
+ * Date          Ticket#  Engineer    Description
+ * ------------- -------- ----------- --------------------------
+ * Apr 23, 2010           bsteffen    Initial creation
+ * Aug 07, 2013  2077     bsteffen    Revise pixel size calculations.
+ * Aug 27, 2013  2287     randerso    Replaced hard coded constant with
+ *                                    densityFactor parameter to allow
+ *                                    application specific density scaling to
+ *                                    better match A1 displays
+ * Sep 10, 2013  16257    MPorricelli Fix so that wind for global grids displays on
+ *                                    mercator maps.
+ * Sep 23, 2013  2363     bsteffen    Add more vector configuration options.
  * 
  * </pre>
  * 
@@ -106,7 +108,7 @@ public abstract class AbstractGriddedDisplay<T> implements IRenderable {
 
     protected IGraphicsTarget target;
 
-    protected final int size;
+    protected final double size;
 
     protected final double densityFactor;
 
@@ -131,7 +133,7 @@ public abstract class AbstractGriddedDisplay<T> implements IRenderable {
      *            adjustment factor to make density match A1
      */
     public AbstractGriddedDisplay(IMapDescriptor descriptor,
-            GeneralGridGeometry gridGeometryOfGrid, int size,
+            GeneralGridGeometry gridGeometryOfGrid, double size,
             double densityFactor) {
         this.calculationQueue = new ConcurrentLinkedQueue<Coordinate>();
